+++ aspects/electrical/resistive/GunnsElectricalResistor.cpp
@@ -177,7 +177,7 @@ void GunnsElectricalResistor::validate() const
     }
 
     /// - Throw an exception if electrical efficiency <= 0 or electrical efficiency >= 1.
-    if (!MsMath::isInRange(static_cast<double>(FLT_EPSILON), mElectricalEfficiency, 1.0 - static_cast<double>(FLT_EPSILON))) {
+    if (!MsMath::isInRange(FLT_EPSILON, static_cast<float>(mElectricalEfficiency), 1.0 - FLT_EPSILON)) {
         GUNNS_ERROR(TsInitializationException, "Invalid Configuration Data", "Electrical Efficiency outside valid range (0-1).");
     }
 }
